backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastapi[standard] >=0.117.1,<1",
    "pydantic-settings >=2.11.0,<3",
]
